Fixed the tests
