+++ spec/Label305/AujaLaravel/Factory/SingleAssociationShowMenuFactorySpec.php
@@ -27,6 +27,7 @@ use Illuminate\Support\Facades\Lang;
 use Illuminate\Support\Facades\URL;
 use Label305\Auja\Menu\LinkMenuItem;
 use Label305\Auja\Menu\Menu;
+use Label305\Auja\Menu\ResourceMenuItem;
 use Label305\Auja\Menu\SpacerMenuItem;
 use Label305\AujaLaravel\Config\Model;
 use Label305\AujaLaravel\Config\Relation;
@@ -34,17 +35,17 @@ use Label305\AujaLaravel\Routing\AujaRouter;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 
-class MultipleAssociationsIndexMenuFactorySpec extends ObjectBehavior {
+class SingleAssociationShowMenuFactorySpec extends ObjectBehavior {
 
     /**
-     * @var Relation[]
+     * @var Relation
      */
-    private $relations;
+    private $relation;
 
     function let(AujaRouter $aujaRouter, Relation $relation, Model $left, Model $right) {
         $this->beConstructedWith($aujaRouter);
 
-        $this->relations = [$relation];
+        $this->relation = $relation;
         $relation->getLeft()->willReturn($left);
         $relation->getRight()->willReturn($right);
         $relation->getType()->willReturn('hasMany');
@@ -55,19 +56,19 @@ class MultipleAssociationsIndexMenuFactorySpec extends ObjectBehavior {
         URL::shouldReceive('route');
         Lang::shouldReceive('trans')->with('Edit')->andReturn('Edit');
         Lang::shouldReceive('trans')->with('Properties')->andReturn('Properties');
-        Lang::shouldReceive('trans')->with('OtherModels')->andReturn('OtherModels');
+        Lang::shouldReceive('trans')->with('OtherModel')->andReturn('OtherModel');
     }
 
     function it_is_initializable() {
-        $this->shouldHaveType('Label305\AujaLaravel\Factory\MultipleAssociationsIndexMenuFactory');
+        $this->shouldHaveType('Label305\AujaLaravel\Factory\SingleAssociationShowMenuFactory');
     }
 
-    function it_can_create_a_menu() {
-        $this->create('Model', 1, [])->shouldHaveType('Label305\Auja\Menu\Menu');
+    function it_can_create_a_menu(Relation $relation) {
+        $this->create('Model', 1, $relation)->shouldHaveType('Label305\Auja\Menu\Menu');
     }
 
     function its_created_menu_should_have_an_edit_linkmenuitem_as_a_first_item() {
-        $menu = $this->create('Model', 1, $this->relations)->getWrappedObject();
+        $menu = $this->create('Model', 1, $this->relation)->getWrappedObject();
         /* @var Menu $menu */
 
         if (!($menu->getMenuItems()[0] instanceof LinkMenuItem)) {
@@ -83,7 +84,7 @@ class MultipleAssociationsIndexMenuFactorySpec extends ObjectBehavior {
     }
 
     function its_created_menu_should_have_a_spacermenuitem_as_a_second_item() {
-        $menu = $this->create('Model', 1, $this->relations)->getWrappedObject();
+        $menu = $this->create('Model', 1, $this->relation)->getWrappedObject();
         /* @var Menu $menu */
 
         if (!($menu->getMenuItems()[1] instanceof SpacerMenuItem)) {
@@ -93,13 +94,13 @@ class MultipleAssociationsIndexMenuFactorySpec extends ObjectBehavior {
         $menuItem = $menu->getMenuItems()[1];
         /* @var $menuItem SpacerMenuItem */
 
-        if ($menuItem->getText() != 'Properties') {
-            throw new \Exception('Text of SpacerMenuItem does equal \'Properties\'');
+        if ($menuItem->getText() != 'OtherModels') {
+            throw new \Exception('Text of SpacerMenuItem does equal \'OtherModels\'');
         }
     }
 
-    function its_created_menu_should_have_model_linkmenuitems() {
-        $menu = $this->create('Model', 1, $this->relations)->getWrappedObject();
+    function its_created_menu_should_have_an_add_model_linkmenuitem() {
+        $menu = $this->create('Model', 1, $this->relation)->getWrappedObject();
         /* @var Menu $menu */
 
         if (!($menu->getMenuItems()[2] instanceof LinkMenuItem)) {
@@ -109,8 +110,17 @@ class MultipleAssociationsIndexMenuFactorySpec extends ObjectBehavior {
         $menuItem = $menu->getMenuItems()[2];
         /* @var $menuItem LinkMenuItem */
 
-        if ($menuItem->getText() != 'OtherModels') {
-            throw new \Exception('Text of LinkMenuItem does equal \'OtherModels\'');
+        if ($menuItem->getText() != 'Add OtherModel') {
+            throw new \Exception('Text of LinkMenuItem does equal \'Add OtherModel\'');
+        }
+    }
+
+    function its_created_menu_should_have_a_resourcemenuitem(){
+        $menu = $this->create('Model', 1, $this->relation)->getWrappedObject();
+        /* @var Menu $menu */
+
+        if (!($menu->getMenuItems()[3] instanceof ResourceMenuItem)) {
+            throw new \Exception('Created Menu has no ResourceMenuItem as a fourth item');
         }
     }
 }
